requirements.txt
Flask 
pyrebase 
python-dotenv
flask-cors